+++ Example/ios/SwiftBridge.swift
@@ -1,0 +1,8 @@
+//
+//  SwiftBridge.swift
+//  RNAudioRecorderPlayer
+//
+//  Created by hyochan on 2021/05/05.
+//
+
+import Foundation